Klarna Loading Placeholder
This is a UIX improvement. 
Klarna loads asynchronously. But the annoying thing is that it makes the page jump when its located between 2 html items that loads during page render.

This solution creates a placeholder with a soft grey to light grey background animation which is seen while Klarna loads. Its an elegant solution to the page jumping